pyproject.toml
[tool.poetry]
name = "lang-graph-app"
version = "0.1.0"
description = ""
authors = ["Quyen Pham Khac <phamkhacquyen1996@gmail.com>"]
readme = "README.md"
packages = [{include = "lang_graph_app"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.3"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.4"
langgraph = "^0.0.64"
langchain-anthropic = "^0.1.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
